services/ingest/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
python-multipart==0.0.9
asyncpg>=0.29.0
google-cloud-storage==2.18.0
anthropic>=0.45.0
openai>=1.50.0
openpyxl==3.1.5
pypdf==4.3.0
python-magic==0.4.27
httpx>=0.27.0
